# backend/openapi/v1.yaml
openapi: 3.0.3
info:
  title: AGS SmartShift API
  version: v1
  description: Sprint 1+ stub — regenerate FE types with `npm run codegen:api` in production/frontend.
paths:
...
  /api/v1/assignments:
    post:
      summary: Assign employee to slot (Shift Leader; dept scope)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [departmentId, slotId, employeeId]
              properties:
                weekId: { type: string }
                departmentId: { type: string, format: uuid }
                slotId: { type: string, format: uuid }
                employeeId: { type: string, format: uuid }
      responses:
        '200':
          description: Assignment created; plan may become in_progress
        '403':
          description: Forbidden (dept scope / role)
...
  /api/v1/assignments/{assignmentId}:
    delete:
      summary: Remove assignment (Shift Leader)
      security:
        - bearerAuth: []
      parameters:
        - name: assignmentId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: weekId
          in: query
          required: true
          schema: { type: string }
      responses:
        '204':
          description: Removed
...
  /api/v1/staffing/days/{weekId}/{dayIdx}:
    get:
      summary: Daily staffing board (PVHK) for a week day
      security:
        - bearerAuth: []
      parameters:
        - { name: weekId, in: path, required: true, schema: { type: string } }
        - { name: dayIdx, in: path, required: true, schema: { type: integer } }
        - { name: departmentCode, in: query, schema: { type: string } }
      responses:
        '200':
          description: Staffing day payload
  /api/v1/staffing/days/{weekId}/{dayIdx}/propose:
    post:
      summary: Propose manning from rules (TBĐH)
      security:
        - bearerAuth: []
      parameters:
        - { name: weekId, in: path, required: true, schema: { type: string } }
        - { name: dayIdx, in: path, required: true, schema: { type: integer } }
        - { name: departmentCode, in: query, schema: { type: string } }
      responses:
        '200':
          description: Updated staffing day
  /api/v1/staffing/days/{weekId}/{dayIdx}/confirm:
    post:
      summary: Confirm assignments and sync ShiftSlot / ShiftAssignment
      security:
        - bearerAuth: []
      parameters:
        - { name: weekId, in: path, required: true, schema: { type: string } }
        - { name: dayIdx, in: path, required: true, schema: { type: integer } }
        - { name: departmentCode, in: query, schema: { type: string } }
      responses:
        '200':
          description: Confirm result
  /api/v1/staffing/weeks/{weekId}/export:
    get:
      summary: Export PVHK week assignment Excel (one sheet per day, template layout)
      security:
        - bearerAuth: []
      parameters:
        - { name: weekId, in: path, required: true, schema: { type: string } }
        - { name: departmentCode, in: query, schema: { type: string } }
      responses:
        '200':
          description: XLSX workbook with 7 day sheets
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
  /api/v1/staffing/days/{weekId}/{dayIdx}/export:
    get:
      summary: Export PVHK daily assignment Excel
      security:
        - bearerAuth: []
      parameters:
        - { name: weekId, in: path, required: true, schema: { type: string } }
        - { name: dayIdx, in: path, required: true, schema: { type: integer } }
        - { name: departmentCode, in: query, schema: { type: string } }
      responses:
        '200':
          description: XLSX file
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
  /api/v1/staffing/days/{weekId}/{dayIdx}/roster:
    get:
      summary: Roster for daily staffing assignment
      security:
        - bearerAuth: []
      parameters:
        - { name: weekId, in: path, required: true, schema: { type: string } }
        - { name: dayIdx, in: path, required: true, schema: { type: integer } }
        - { name: departmentCode, in: query, schema: { type: string } }
      responses:
        '200':
          description: Roster entries
  /api/v1/staffing/days/{weekId}/{dayIdx}/auto-assign:
    post:
      summary: Auto-assign crew proposals (shift leader)
      security:
        - bearerAuth: []
      parameters:
        - { name: weekId, in: path, required: true, schema: { type: string } }
        - { name: dayIdx, in: path, required: true, schema: { type: integer } }
        - { name: departmentCode, in: query, schema: { type: string } }
      responses:
        '200':
          description: Auto-assign result with proposals and warnings
  /api/v1/staffing/days/{weekId}/{dayIdx}/publish:
    post:
      summary: Publish and lock staffing day (push to staff)
      security:
        - bearerAuth: []
      parameters:
        - { name: weekId, in: path, required: true, schema: { type: string } }
        - { name: dayIdx, in: path, required: true, schema: { type: integer } }
        - { name: departmentCode, in: query, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                lockReason: { type: string, nullable: true }
      responses:
        '200':
          description: Published
  /api/v1/staffing/days/{weekId}/{dayIdx}/bio-header:
    get:
      summary: BIO header for staffing day
      security:
        - bearerAuth: []
      parameters:
        - { name: weekId, in: path, required: true, schema: { type: string } }
        - { name: dayIdx, in: path, required: true, schema: { type: integer } }
        - { name: departmentCode, in: query, schema: { type: string } }
      responses:
        '200':
          description: BIO header
    put:
      summary: Update BIO header
      security:
        - bearerAuth: []
      parameters:
        - { name: weekId, in: path, required: true, schema: { type: string } }
        - { name: dayIdx, in: path, required: true, schema: { type: integer } }
        - { name: departmentCode, in: query, schema: { type: string } }
      responses:
        '200':
          description: Updated BIO header
  /api/v1/staffing/assignments:
    post:
      summary: Create flight crew assignment on staffing line
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Assignment created
  /api/v1/staffing/assignments/{id}:
    patch:
      summary: Update assignment work window (shift leader calendar)
      security:
        - bearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Updated assignment
    delete:
      summary: Remove staffing assignment
      security:
        - bearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '204':
          description: Deleted
  /api/v1/staffing/lines/{id}:
    patch:
      summary: Patch target manning (TBĐH)
      security:
        - bearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Updated line
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT